# be/api/ollamahandler.go
package api

import (
	"bytes"
	"encoding/json"
	"fmt"
	"io"
	"log"
	"net/http"
	"rprj/be/dblayer"
	"strings"
	"time"
)

var ollamaURL string
var ollamaModel string
var ollamaAppName string

var ollamaFolder *dblayer.DBFolder

func OllamaInit(appName, url, model string) error {
	ollamaURL = url
	ollamaModel = model
	ollamaAppName = appName

	if ollamaURL != "" && ollamaModel != "" {
		// go UpdateOllamaDefaultPageResponse("en")
		log.Printf("Ollama initialized with URL: %s and Model: %s\n", ollamaURL, ollamaModel)

		OllamaFolderInit("Ollama Pages")

	} else {
		log.Println("Ollama not configured - using fallback responses")
	}
	lastDefaultPageResponse = "<h2>Welcome! 👋</h2><p>Please log in to continue using the application.</p>"

	return nil
}

func OllamaFolderInit(folderName string) {
	dbContext := &dblayer.DBContext{
		UserID:   "-1",           // DANGEROUS!!!! Think of something better here!!!
		GroupIDs: []string{"-2"}, // Same here!!!
		Schema:   dblayer.DbSchema,
	}

	repo := dblayer.NewDBRepository(dbContext, dblayer.Factory, dblayer.DbConnection)
	repo.Verbose = false

	search := repo.GetInstanceByTableName("folders")
	if search == nil {
		log.Println("Failed to create folder instance for Ollama")
		return
	}
	search.SetValue("name", folderName)
	results, err := repo.Search(search, false, false, "")
	if err != nil {
		log.Printf("Failed to find or create Ollama folder '%s': %v\n", folderName, err)
		return
	}
	if len(results) == 1 {
		ollamaFolder = results[0].(*dblayer.DBFolder)
		return
	}

	// Create the folder
	newFolder := repo.GetInstanceByTableName("folders")
	if newFolder == nil {
		log.Println("Failed to create new folder instance for Ollama")
		return
	}
	newFolder.SetValue("name", folderName)
	newFolder.SetValue("description", "AI-generated content by Ollama")
	newFolder.SetValue("permissions", "rwxr--r--") // Tutti possono leggere
	created, err := repo.Insert(newFolder)
	if err != nil {
		log.Printf("Failed to create Ollama folder '%s': %v\n", folderName, err)
		return
	}
	ollamaFolder = created.(*dblayer.DBFolder)
	log.Printf("Created Ollama folder '%s' with ID %s\n", folderName, ollamaFolder.GetValue("id"))
}

// OllamaHandler godoc
//
//	@Summary sends a prompt to Ollama and returns the response
//	@Description Sends a prompt to the Ollama API and returns the generated response
//	@Tags ollama
//	@Accept json
//	@Produce json
//	@Param request body map[string]string true "Request body containing the prompt"
//	@Success 200 {object} map[string]string "Ollama response"
//	@Failure 400 {object} map[string]string "Invalid request"
//	@Failure 503 {object} map[string]string "Ollama service not configured"
//	@Router /ollama [post]
func OllamaHandler(w http.ResponseWriter, r *http.Request) {
	type Request struct {
		Prompt string `json:"prompt"`
	}
	type Response struct {
		Response string `json:"response"`
		Error    string `json:"error,omitempty"`
	}

	var req Request
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		RespondSimpleError(w, ErrInvalidRequest, "Invalid request body", http.StatusBadRequest)
		return
	}

	if ollamaURL == "" || ollamaModel == "" {
		res := Response{Error: "Ollama service not configured"}
		w.WriteHeader(http.StatusServiceUnavailable)
		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(res)
		return
	}

	respText, err := CallOllama(req.Prompt)
	if err != nil {
		res := Response{Error: err.Error()}
		w.Header().Set("Content-Type", "application/json")
		json.NewEncoder(w).Encode(res)
		return
	}

	res := Response{Response: respText}
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(res)
}

var lastDefaultPageResponse string

// DefaultPageOllamaHandler godoc
//
//	@Summary returns default page response generated by Ollama
//	@Description Returns the default page response generated by Ollama in the requested language
//	@Tags ollama
//	@Produce json
//	@Param lang query string false "Language tag (e.g., en, it, fr, de)" default(en)
//	@Success 200 {object} map[string]string "Default page response"
//	@Failure 503 {object} map[string]string "Ollama service not configured"
//	@Router /ollama/default-page [get]
func DefaultPageOllamaHandler(w http.ResponseWriter, r *http.Request) {
	// Get lang query parameter
	langParam := r.URL.Query().Get("lang")
	if langParam == "" {
		langParam = "en"
	}
	log.Print("Requested language tag: ", langParam)

	// Just use English for now
	langParam = "en"
	go UpdateOllamaDefaultPageResponse(langParam)

	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"response": lastDefaultPageResponse,
	})
}

var prompts = []string{
	"Provide a short funny star trek based welcome message for the default page of ",
	"Provide a short funny star wars based welcome message for the default page of ",
	"Provide a short funny hitcher's guide to galaxy based welcome message for the default page of ",
	"Provide a short funny positive welcome message for the default page of ",
}

func UpdateOllamaDefaultPageResponse(languageTag string) {
	if ollamaURL == "" || ollamaModel == "" {
		log.Print("Ollama not configured, skipping default page response update.")
		return
	}

	// ollamaModel = modelList[0]
	prompt := prompts[2]
	prompt += " a web application"
	// prompt += " " + ollamaAppName + " web application"
	switch languageTag {
	case "it", "it-IT":
		prompt += prompt + " in Italian."
	case "en", "en-US":
		prompt += prompt + " in English."
	case "fr", "fr-FR":
		prompt += prompt + " in French."
	case "de", "de-DE":
		prompt += prompt + " in German."
	}
	prompt += " Use emojis."
	prompt += " Use only <h2>, <p>, <b>, <br> tags. No links <a>. Output html code only."

	// prompt = "Créez un court message de bienvenue humoristique, "
	// prompt += "inspiré de animal crossing, "
	// prompt += "pour la page d'accueil d'une application web. "
	// prompt += "Rédigez ce message en français. "
	// prompt += "Utilisez des émojis. "
	// prompt += "Utilisez uniquement les tags <h2>, <p>, <b> et <br>. Pas de liens <a>. Générez uniquement du code HTML."

	log.Print("Prompt: ", prompt)

	respText, err := CallOllama(prompt)
	if err != nil {
		log.Printf("Error getting default page response from Ollama: %v\n", err)
		respText = "Welcome to our application!"
	}

	lastDefaultPageResponse = respText

	OllamaSavePage(respText)
}

func OllamaSavePage(content string) {
	if ollamaFolder == nil {
		log.Println("Ollama folder not initialized, cannot save page.")
		return
	}

	dbContext := &dblayer.DBContext{
		UserID:   "-1",           // DANGEROUS!!!! Think of something better here!!!
		GroupIDs: []string{"-2"}, // Same here!!!
		Schema:   dblayer.DbSchema,
	}

	repo := dblayer.NewDBRepository(dbContext, dblayer.Factory, dblayer.DbConnection)
	repo.Verbose = false

	page := repo.GetInstanceByTableName("pages")
	if page == nil {
		log.Println("Failed to create page instance for Ollama")
		return
	}
	// Generate title with YYYY.MM.DD prefix
	titlePrefix := time.Now().Format("2006.01.02 15:04")
	page.SetValue("name", titlePrefix+" Ollama Default Page")
	page.SetValue("description", "AI-generated content by Ollama")
	page.SetValue("html", content)
	page.SetValue("father_id", ollamaFolder.GetValue("id"))
	page.SetValue("permissions", "rwxrw-r--") // Tutti possono leggere
	page.SetValue("group_id", "-6")
	created, err := repo.Insert(page)
	if err != nil {
		log.Printf("Failed to create Ollama default page: %v\n", err)
		return
	}
	createdPage := created.(*dblayer.DBPage)
	log.Printf("Created Ollama default page with ID %s\n", createdPage.GetValue("id"))
}

// CallOllama sends a prompt to the Ollama API and returns the response
func CallOllama(prompt string) (string, error) {
	if ollamaURL == "" || ollamaModel == "" {
		return "", fmt.Errorf("Ollama not configured")
	}

	requestBody, err := json.Marshal(map[string]interface{}{
		"model": ollamaModel,
		"messages": []map[string]string{
			{
				"role":    "user",
				"content": prompt,
			},
		},
		"stream": false,
	})
	if err != nil {
		return "", err
	}

	resp, err := http.Post(fmt.Sprintf("%s", ollamaURL), "application/json", bytes.NewBuffer(requestBody))
	if err != nil {
		return "", err
	}
	defer resp.Body.Close()

	body, err := io.ReadAll(resp.Body)
	if err != nil {
		return "", err
	}

	log.Printf("Ollama response: %s\n", string(body))

	// Parse the response
	// {"model":"llama3.2:latest","created_at":"2025-11-13T13:17:45.487900903Z","message":{"role":"assistant","content":"I'm an artificial intelligence model known as Llama. Llama stands for \"Large Language Model Meta AI.\""},"done":true,"done_reason":"stop","total_duration":1593273648,"load_duration":103413131,"prompt_eval_count":29,"prompt_eval_duration":132183696,"eval_count":23,"eval_duration":1338392262}

	var ollamaResponseSingle struct {
		Message struct {
			Content string `json:"content"`
		} `json:"message"`
	}

	if err := json.Unmarshal(body, &ollamaResponseSingle); err == nil && ollamaResponseSingle.Message.Content != "" {
		content := ollamaResponseSingle.Message.Content
		content = strings.ReplaceAll(content, "```html", "")
		content = strings.ReplaceAll(content, "```", "")
		content = strings.ReplaceAll(content, "[app name]", ollamaAppName)
		content = strings.ReplaceAll(content, "[App Name]", ollamaAppName)
		return content, nil
	}

	return "", fmt.Errorf("invalid response from Ollama")

}