# be/docs/swagger.yaml
basePath: /
...
host: localhost:1971
info:
  contact:
    email: support@example.com
    name: API Support
  description: API documentation for R-PRJ-NG backend
  license:
    name: Apache 2.0
    url: https://opensource.org/licenses/Apache-2.0
  termsOfService: http://swagger.io/terms/
  title: ρBee (rhobee) API
  version: "1.0"
paths:
...
  /ollama:
    post:
      consumes:
      - application/json
      description: Sends a prompt to the Ollama API and returns the generated response
      parameters:
      - description: Request body containing the prompt
        in: body
        name: request
        required: true
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Ollama response
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Ollama service not configured
          schema:
            additionalProperties:
              type: string
            type: object
      summary: sends a prompt to Ollama and returns the response
      tags:
      - ollama
  /ollama/default-page:
    get:
      description: Returns the default page response generated by Ollama in the requested
        language
      parameters:
      - default: en
        description: Language tag (e.g., en, it, fr, de)
        in: query
        name: lang
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Default page response
          schema:
            additionalProperties:
              type: string
            type: object
        "503":
          description: Ollama service not configured
          schema:
            additionalProperties:
              type: string
            type: object
      summary: returns default page response generated by Ollama
      tags:
      - ollama
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...